requirements.txt
nxt-python
PyUSB
